+++ frontend/src/components/organisms/ContactModal/ContactModal.jsx
@@ -4,10 +4,10 @@ import PropTypes from 'prop-types'
 import { faTimes } from '@fortawesome/free-solid-svg-icons'
 import { useForm } from 'react-hook-form'
 import { yupResolver } from '@hookform/resolvers/yup'
-import Button from '../../atoms/Button/Button'
+import Button from '../../atoms/Forms/Button'
 import Modal from '../Modal/Modal'
 import Input from '../../atoms/Input/Input'
-import TextArea from '../../atoms/TextArea/TextArea'
+import TextArea from '../../atoms/Forms/TextArea'
 
 import { ButtonWrapper } from './ContactModal.style'
 import { colors } from '../../../theme'
