+++ frontend/src/components/organisms/DeleteModal/DeleteModal.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react'
 import PropTypes from 'prop-types'
 import { faTimes } from '@fortawesome/free-solid-svg-icons'
-import Button from '../../atoms/Button/Button'
+import Button from '../../atoms/Forms/Button'
 import Modal from '../Modal/Modal'
 import Input from '../../atoms/Input/Input'
 import useInput from '../../../hooks/useInput'
