+++ frontend/src/components/organisms/EditProfileModal/EditProfile.jsx
@@ -2,7 +2,7 @@
 import React, { useState } from 'react'
 import PropTypes from 'prop-types'
 import { faTimes } from '@fortawesome/free-solid-svg-icons'
-import Button from '../../atoms/Button/Button'
+import Button from '../../atoms/Forms/Button'
 import Modal from '../Modal/Modal'
 import Input from '../../atoms/Input/Input'
 // STYLES
