+++ frontend/src/components/organisms/UserModal/UserModal.jsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react'
 import PropTypes from 'prop-types'
 import { faTimes } from '@fortawesome/free-solid-svg-icons'
-import Button from '../../atoms/Button/Button'
+import Button from '../../atoms/Forms/Button'
 import Modal from '../Modal/Modal'
 import { UserModalStyled, ButtonWrapper } from './UserModal.style'
 import Colors from '../../../theme/colors'
