+++ frontend/src/pages/AdList/AdCard/AdCard.jsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react'
 import PropTypes from 'prop-types'
 import { useHistory } from 'react-router-dom'
 import { faComments } from '@fortawesome/free-solid-svg-icons'
-import Button from '../../../components/atoms/Button/Button'
+import Button from '../../../components/atoms/Forms/Button'
 import { colors } from '../../../theme'
 import ContactModal from '../../../components/organisms/ContactModal/ContactModal'
 import casaPiscinaAd from '../../../assets/images/casaPiscinaAd2.jpg'
