+++ frontend/src/pages/AdList/AdList.jsx
@@ -4,7 +4,7 @@ import axios from 'axios'
 import _ from 'lodash'
 import AdCard from './AdCard/AdCard'
 import Body from '../../components/layout/Body/Body'
-import Button from '../../components/atoms/Button/Button'
+import Button from '../../components/atoms/Forms/Button'
 import Colors from '../../theme/colors'
 import MapView from '../../components/organisms/Map/MapView/MapView'
 import AdListFilter from './AdListFilter/AdListFilter'
