+++ frontend/src/pages/CreateNewAd/CreateNewAd.jsx
@@ -14,8 +14,8 @@ import newAdSchema from '../../validation/createNewAdSchema'
 import { newNotification, NotificationTypes } from '../../store/notificationSlice'
 import Body from '../../components/layout/Body/Body'
 import InputNumber from '../../components/atoms/InputNumber/InputNumber'
-import TextArea from '../../components/atoms/TextArea/TextArea'
-import Button from '../../components/atoms/Button/Button'
+import TextArea from '../../components/atoms/Forms/TextArea'
+import Button from '../../components/atoms/Forms/Button'
 import Input from '../../components/atoms/Input/Input'
 import Modal from '../../components/organisms/Modal/Modal'
 import {
