+++ frontend/src/pages/EditAd/EditAd.jsx
@@ -1,10 +1,5 @@
 import { useState } from 'react'
 import { useHistory } from 'react-router-dom'
-import Body from 'components/layout/Body/Body'
-import Input from 'components/atoms/Input/Input'
-import InputNumber from 'components/atoms/InputNumber/InputNumber'
-import TextArea from 'components/atoms/TextArea/TextArea'
-import Button from 'components/atoms/Button/Button'
 import {
   faMapMarkerAlt,
   faBed,
@@ -14,9 +9,14 @@ import {
 } from '@fortawesome/free-solid-svg-icons'
 
 // Styles
+import Container from '../../theme/globalStyles'
+import CustomMap from '../../components/organisms/Map/CustomMap/CustomMap'
+import Button from '../../components/atoms/Forms/Button'
+import InputNumber from '../../components/atoms/InputNumber/InputNumber'
+import Input from '../../components/atoms/Input/Input'
+import Body from '../../components/layout/Body/Body'
 import { Wrapper, MapText } from './EditAd.styles'
-import { Container } from 'theme/GlobalStyles'
-import CustomMap from 'components/organisms/Map/CustomMap'
+import TextArea from '../../components/atoms/Forms/TextArea'
 
 const EditAd = (props) => {
   const originalAd = Object.assign({}, props.location.state.ad)
