+++ frontend/src/pages/UserFlow/Login/Login.jsx
@@ -5,7 +5,7 @@ import { useForm } from 'react-hook-form'
 import { yupResolver } from '@hookform/resolvers/yup'
 import { useDispatch } from 'react-redux'
 import Body from '../../../components/layout/Body/Body'
-import AsyncButton from '../../../components/atoms/Button/Button'
+import AsyncButton from '../../../components/atoms/Forms/Button'
 import { Container, Form, RedirectStyled } from '../UserFlow.styles'
 import { InputGroup } from '../../../components/molecules'
 import loginSchema from '../../../validation/loginSchema'
