+++ frontend/src/pages/UserFlow/Registration/Registration.jsx
@@ -10,13 +10,13 @@ import Body from '../../../components/layout/Body/Body'
 
 // Units Components
 import CheckBox from '../../../components/atoms/CheckBox'
-import AsyncButton from '../../../components/atoms/Button/Button'
+import AsyncButton from '../../../components/atoms/Forms/Button'
 
 // Styles
 import { Container, Form, RedirectStyled } from '../UserFlow.styles'
 
 // Utilities
-import Input from '../../../components/molecules/InputGroup'
+import InputGroup from '../../../components/molecules/InputGroup'
 
 // eslint-disable-next-line import/extensions
 import registerSchema from '../../../validation/registerUserSchema.js'
@@ -110,7 +110,8 @@ function Register() {
     <Body title="Registro" justifyTitle="center">
       <Container>
         <Form onSubmit={handleSubmit(submitForm)} noValidate>
-          <Input
+          <InputGroup
+            label="text"
             type="text"
             placeholder="Nombre"
             id="name"
@@ -121,7 +122,8 @@ function Register() {
             register={register('name')}
           />
 
-          <Input
+          <InputGroup
+            label="text"
             type="text"
             placeholder="Apellido"
             id="lastname"
@@ -132,7 +134,8 @@ function Register() {
             register={register('lastname')}
           />
 
-          <Input
+          <InputGroup
+            label="email"
             type="email"
             placeholder="Email"
             id="emailName"
@@ -143,7 +146,8 @@ function Register() {
             register={register('email')}
           />
 
-          <Input
+          <InputGroup
+            label="password"
             type="password"
             placeholder="Contraseña"
             id="password"
